Moved the majority of types into src/types folder
