+++ src/api/sakApi.ts
@@ -1,13 +1,6 @@
-import apiClient, { ApiClientResult } from './apiClient';
-import { Behandling } from './behandlingApi';
-import { Søknad } from './søknadApi';
+import { Sak } from '~types/Sak';
 
-export interface Sak {
-    id: string;
-    fnr: string;
-    behandlinger: Behandling[];
-    søknader: Array<Søknad>;
-}
+import apiClient, { ApiClientResult } from './apiClient';
 
 export async function fetchSakByFnr(fnr: string): Promise<ApiClientResult<Sak>> {
     return apiClient({ url: `/saker?fnr=${fnr}`, method: 'GET' });
