+++ src/pages/saksoversikt/simulering/simulering.tsx
@@ -7,12 +7,12 @@ import { Innholdstittel } from 'nav-frontend-typografi';
 import React from 'react';
 
 import messages from '~/features/beregning/beregning-nb';
-import { Utbetaling } from '~api/behandlingApi';
-import { Sak } from '~api/sakApi';
 import { formatDateTime } from '~lib/dateUtils';
 import { combineOptions } from '~lib/fp';
 import { useI18n } from '~lib/hooks';
 import { useAppSelector } from '~redux/Store';
+import { Sak } from '~types/Sak';
+import { Utbetaling } from '~types/Utbetaling';
 
 import styles from '../beregning/visBeregning.module.less';
 import { groupSimuleringsperioder } from '../delt/arrayUtils';
