+++ src/pages/saksoversikt/vilkår/Sats.tsx
@@ -3,7 +3,7 @@ import { Textarea } from 'nav-frontend-skjema';
 import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 
-import { Sats as FaktiskSats } from '~api/behandlingApi';
+import { Sats as FaktisktSats } from '~/types/Sats';
 import { SuperRadioGruppe } from '~components/FormElements';
 import { lagreBehandlingsinformasjon } from '~features/saksoversikt/sak.slice';
 import { DelerBoligMed } from '~features/søknad/types';
@@ -42,14 +42,14 @@ const utledSats = (values: FormData) => {
         return null;
     }
     if (!values.delerSøkerBolig) {
-        return FaktiskSats.Høy;
+        return FaktisktSats.Høy;
     }
     switch (values.delerBoligMedHvem) {
         case null:
             return null;
         case DelerBoligMed.VOKSNE_BARN:
         case DelerBoligMed.ANNEN_VOKSEN:
-            return FaktiskSats.Lav;
+            return FaktisktSats.Lav;
         case DelerBoligMed.EKTEMAKE_SAMBOER:
             switch (values.erEktemakeEllerSamboerUnder67) {
                 case null:
