+++ src/types/Beregning.ts
@@ -1,0 +1,22 @@
+import { Fradrag } from './Fradrag';
+import { Sats } from './Sats';
+
+export interface Beregning {
+    id: string;
+    opprettet: string;
+    sats: Sats;
+    fom: string;
+    tom: string;
+    månedsberegninger: Array<Månedsberegning>;
+    fradrag: Array<Fradrag>;
+}
+
+export interface Månedsberegning {
+    id: string;
+    sats: Sats;
+    beløp: number;
+    grunnbeløp: number;
+    fom: string;
+    tom: string;
+    fradrag: number;
+}
