+++ src/types/Fradrag.ts
@@ -1,0 +1,16 @@
+import { Nullable } from '~lib/types';
+
+export interface Fradrag {
+    type: Fradragstype;
+    beløp: number;
+    beskrivelse: Nullable<string>;
+}
+
+export enum Fradragstype {
+    Uføretrygd = 'Uføretrygd',
+    Barnetillegg = 'Barnetillegg',
+    Arbeidsinntekt = 'Arbeidsinntekt',
+    Pensjon = 'Pensjon',
+    Kapitalinntekt = 'Kapitalinntekt',
+    AndreYtelser = 'AndreYtelser',
+}
