+++ src/types/Sak.ts
@@ -1,0 +1,9 @@
+import { Behandling } from './Behandling';
+import { Søknad } from './Søknad';
+
+export interface Sak {
+    id: string;
+    fnr: string;
+    behandlinger: Behandling[];
+    søknader: Array<Søknad>;
+}
