+++ src/types/Sats.ts
@@ -1,0 +1,4 @@
+export enum Sats {
+    Høy = 'HØY',
+    Lav = 'LAV',
+}
