+++ src/types/Simulering.ts
@@ -1,0 +1,10 @@
+export interface Simulering {
+    totalBruttoYtelse: number;
+    perioder: SimulertPeriode[];
+}
+
+export interface SimulertPeriode {
+    fom: string;
+    tom: string;
+    bruttoYtelse: number;
+}
