+++ src/types/Utbetaling.ts
@@ -1,0 +1,7 @@
+import { Simulering } from './Simulering';
+
+export interface Utbetaling {
+    id: string;
+    opprettet: string;
+    simulering: Simulering;
+}
